this segfaulted on s390

+++ loader/loader.c
@@ -464,7 +464,7 @@ static int setupStage2Image(int fd, char * dest, int flags,
 
     rc = copyFileFd(fd, dest);
     stat(dest, &sb);
-    logMessage("copied %d bytes to %s", sb.st_size, dest);
+    logMessage("copied %ld bytes to %s", (long)sb.st_size, dest);
 
     if (rc) {
 	/* just to make sure */
